More refactoring to make the mobile view work nicely
