+++ components/PlayerPage/Details/PlayerInfo.tsx
